backend/pyproject.toml
[project]
name = "gnn-imbalance-backend"
version = "1.0.0"
description = "Backend API for Explainable GNN Imbalance Studio"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.9.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.20.0",
    "psycopg2-binary>=2.9.0",
    "redis>=5.0.0",
    "celery>=5.4.0",
    "torch>=2.2.0",
    "torch_geometric>=2.5.0",
    "torch-scatter>=2.1.0",
    "torch-sparse>=0.6.0",
    "scikit-learn>=1.5.0",
    "numpy>=1.26.0",
    "networkx>=3.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.backends"
